+++ src/_data/pages/index.js
@@ -2,6 +2,7 @@ import i18n from '../i18n.js'
 import { getGuidePageDefinition } from '../guide-pages.js'
 import home from './home.js'
 import registration from './registration.js'
+import revokeRegistration from './revoke-registration.js'
 import contact from './contact.js'
 import faq from './faq.js'
 import stats from './stats.js'
@@ -13,6 +14,7 @@ const pageDefinitionsByPath = {
     '/faq/':                          faq,
     '/contact/':                      contact,
     '/registration/':                 registration,
+    '/registration/revoke/':          revokeRegistration,
     '/stats/':                        stats,
     '/changelog/':                    changelog,
     '/licensing/':                    legal.licensing,
@@ -33,6 +35,7 @@ export default {
     home,
     contact,
     registration,
+    revokeRegistration,
     faq,
     legal,
     changelog,